Design code smell - 1 - Refactored code for Insufficient Modularization using extract class technique
